feat(suite): ETH tokens moved to own tab next to Overview tab
