+++ packages/suite-web/pages/accounts/tokens/index.tsx
@@ -1,0 +1,3 @@
+import Tokens from '@wallet-views/tokens';
+
+export default Tokens;
