+++ packages/suite/src/components/suite/AppNavigation/index.tsx
@@ -148,7 +148,8 @@ const isSubsection = (routeName: Route['name']): boolean =>
     !(
         routeName.startsWith('settings') ||
         routeName === 'wallet-index' ||
-        routeName === 'wallet-details'
+        routeName === 'wallet-details' ||
+        routeName === 'wallet-tokens'
     );
 
 const isSecondaryMenuOverflown = ({ primary, secondary, wrapper }: MenuWidths) =>
