+++ packages/suite/src/constants/suite/routes.ts
@@ -213,6 +213,12 @@ const routes = [
         params: walletParams,
     },
     {
+        name: 'wallet-tokens',
+        pattern: '/accounts/tokens',
+        app: 'wallet',
+        params: walletParams,
+    },
+    {
         name: 'passwords-index',
         pattern: '/passwords',
         app: 'passwords',
